+++ frontend/src/App.tsx
@@ -3,16 +3,17 @@ import { BrowserRouter } from 'react-router-dom';
 
 import { QueryClientProvider } from '@tanstack/react-query';
 
+import queryClient from 'api/config/queryClient';
+import UserAgentProvider from 'common/contexts/UserAgent';
+import ErrorPage from 'service/pages/ErrorPage';
+import 'styles/@app.scss';
+
 import { RecoilRoot } from 'recoil';
 import PageRoutes from 'routes';
 
-import { ErrorBoundary, SnackbarProvider } from 'common/components';
+import { ErrorBoundary, Snackbar } from 'common/components';
 
-import queryClient from 'api/config/queryClient';
-import UserAgentProvider from 'common/contexts/UserAgent';
 import ModalProvider from 'service/components/ModalProvider';
-import ErrorPage from 'service/pages/ErrorPage';
-import 'styles/@app.scss';
 
 function ContextWrapper({ children }: { children: ReactNode }) {
   return (
@@ -30,7 +31,7 @@ function ContextWrapper({ children }: { children: ReactNode }) {
 function App() {
   return (
     <ContextWrapper>
-      <SnackbarProvider />
+      <Snackbar.Provider />
       <Suspense>
         <ErrorBoundary fallback={ErrorPage}>
           <PageRoutes />
